concat2's hw-emu fails, other kernels are disabled in CMakeLists for fast build, CMakeList kernel cflag bug is fixed

+++ inc/ocl_imp/xilinx/VectorizationHelper.h
@@ -11,8 +11,8 @@ struct VectorizedArray{
 };
 
 // Be careful with parenthesis in a preprocessor, especially when the expression has math ops!
-#define FlatIdx_to_VecIdx(vecDepth, flatIdx) ((flatIdx)/(vecDepth))
-#define FlatIdx_to_VecSubIdx(vecDepth, flatIdx) ((flatIdx)%(vecDepth))
+#define FlatIdx_to_VecIdx(vecDepth, flatIdx) (((unsigned long)flatIdx)/((unsigned long)vecDepth))
+#define FlatIdx_to_VecSubIdx(vecDepth, flatIdx) (((unsigned long)flatIdx)%((unsigned long)vecDepth))
 
 // Input and output tensor vectorization depth for each of the kernels:(words of 4-bytes)
 #define CONFIG_M_AXI_WIDTH	16
